restructure project & ad support for file watch & abstract backends
